Added room passwords (optional)
You can choose to put in a room password. Allows for the ClientRunner to handle how the room name and password are gotten from the user. Has to be sent to the server in roomName:roomPassword format, otherwise, it won't work. You can choose to not have a password for the room by just sending an empty string after the colon (":")