7 server code (#28)
* Imported Server Code

Fixed the errors with the imported server code from my eclipse project.

* Added rooms to join

* Added room passwords (optional)

You can choose to put in a room password. Allows for the ClientRunner to handle how the room name and password are gotten from the user. Has to be sent to the server in roomName:roomPassword format, otherwise, it won't work. You can choose to not have a password for the room by just sending an empty string after the colon (":")

* Added Authentication!

* Changed to Executor Threading, and added javadocs

* Message length and dense packets

* Added States

* Added ability to get all states of a type

---------

Co-authored-by: Swordpartee <117320861+Swordpartee@users.noreply.github.com>